cloud_function/requirements.txt
pandas==2.2.2
google-cloud-aiplatform==1.49.0
google-cloud-storage==2.16.0
pyarrow==16.1.0
functions-framework==3.5.0
gcsfs==2023.12.2.post1
google-generativeai==0.7.1